+++ api/CHANGELOG.md
@@ -63,6 +63,7 @@
 - Added a `Tools` tab
 - Added a `Merge split APKs` tool in the `Tools` tab that just merges the selected split APK and allows the user to save it to storage after https://github.com/Jman-Github/Universal-ReVanced-Manager/issues/171
 - Added a `Create custom YouTube icons & headers` tool to the `Tools` tab (inspired by [Morphe Managers implementation](https://github.com/MorpheApp/morphe-manager/pull/138))
+- Made main screen tab titles wrap to prevent them from being cut  off https://github.com/Jman-Github/Universal-ReVanced-Manager/issues/175
 
 
 # Bug fixes